requirements.txt
# requirements.txt
requests
psycopg2-binary
tenacity
python-dotenv
dataclasses-json